USD 22.05b and slang upgrade (#184)
* win fixes + slang v0.23.14 + usd 22.05b upgrade

* pick discrete gpu

* build instruction for usd

+++ include/settings/settings.h
@@ -50,7 +50,7 @@ public:
     float getAs(const char* name)
     {
         isNameValid(name);
-        return atof(mMap[name].c_str());
+        return (float) atof(mMap[name].c_str());
     }
 
     template <>
